Added button to expanded QR code view
