+++ .idea/misc.xml
@@ -3,8 +3,6 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="../../../../layout/custom_preview.xml" value="0.2677083333333333" />
-        <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_other_profile.xml" value="0.33" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.20531526836894215" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/colorgradient.xml" value="0.3828125" />
         <entry key="..\:/Users/Dallin Dmytryk/Desktop/Git/QR-Scape/code/app/src/main/res/drawable/edit_text_background.xml" value="0.3828125" />
@@ -16,6 +14,8 @@
         <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_personal_stats.xml" value="0.1286231884057971" />
         <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/activity_scan_view.xml" value="0.1286231884057971" />
         <entry key="..\:/Users/Dallin/Desktop/GIT/QR-Scape/app/src/main/res/layout/toolbar_title_layout.xml" value="0.1286231884057971" />
+        <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_expanded_qr_view.xml" value="0.2796875" />
+        <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_other_profile.xml" value="0.25" />
         <entry key="..\:/Users/Owner/AndroidStudioProjects/QR-Scape/app/src/main/res/layout/activity_profile.xml" value="0.25" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_home.xml" value="0.12454710144927536" />
         <entry key="..\:/Users/harsh/QR-Scape/app/src/main/res/layout/activity_leaderboard.xml" value="0.30833333333333335" />
