+++ app/src/main/java/com/example/qr_scape/Expanded_QR_View.java
@@ -1,10 +1,18 @@
 package com.example.qr_scape;
 
+import static android.content.ContentValues.TAG;
+
 import android.content.Intent;
 import android.os.Bundle;
+import android.util.Log;
 import android.widget.TextView;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.firebase.firestore.FirebaseFirestore;
 
 public class Expanded_QR_View extends AppCompatActivity {
 
@@ -35,4 +43,35 @@ public class Expanded_QR_View extends AppCompatActivity {
 
 
     }
+
+    /**
+     * Deletes an instance of a user QR code from the database
+     * @param qrCode
+     * @author Ty Greve
+     * @version 1
+     */
+    public void deleteQRCode(QRCode qrCode){
+        // Deletes an instance (scan by a user) of a QR code. QRCode (real/physical) remains in the database
+
+        // Access a Cloud Firestore instance from your Activity
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+
+        // Delete the document from the Firestore QRCodeInstance Collection
+        // Get reference to Firestore collection and Document ID
+        db.collection("QRCodeInstance").document(qrCode.getQRHashSalted())
+                .delete() // delete document in the Firestore database
+                .addOnSuccessListener(new OnSuccessListener<Void>() {
+                    @Override
+                    public void onSuccess(Void aVoid) {
+                        Log.d(TAG, "Document has been deleted successfully!");
+                    }
+                })
+                .addOnFailureListener(new OnFailureListener() {
+                    @Override
+                    public void onFailure(@NonNull Exception e) {
+                        Log.d(TAG, "Error deleting the document!" + e.toString());
+                    }
+                });
+    }//end deleteQRCode
+
 }
