+++ app/src/main/java/com/example/qr_scape/QRCollectionActivity.java
@@ -100,37 +100,6 @@ public class QRCollectionActivity extends AppCompatActivity {
         });
     }
 
-
-    /**
-     * Deletes an instance of a user QR code from the database
-     * @param qrCode
-     * @author Ty Greve
-     * @version 1
-     */
-    public void deleteQRCode(QRCode qrCode){
-        // Deletes an instance (scan by a user) of a QR code. QRCode (real/physical) remains in the database
-
-        // Access a Cloud Firestore instance from your Activity
-        FirebaseFirestore db = FirebaseFirestore.getInstance();
-
-        // Delete the document from the Firestore QRCodeInstance Collection
-        // Get reference to Firestore collection and Document ID
-        db.collection("QRCodeInstance").document(qrCode.getQRHashSalted())
-                .delete() // delete document in the Firestore database
-                .addOnSuccessListener(new OnSuccessListener<Void>() {
-                    @Override
-                    public void onSuccess(Void aVoid) {
-                        Log.d(TAG, "Document has been deleted successfully!");
-                    }
-                })
-                .addOnFailureListener(new OnFailureListener() {
-                    @Override
-                    public void onFailure(@NonNull Exception e) {
-                        Log.d(TAG, "Error deleting the document!" + e.toString());
-                    }
-                });
-    }//end deleteQRCode
-
     /**
      * Delete every instance of user QR codes (scanned instances) and the
      * real/physical QR code from the database
